Migrate ways of working content
This takes the published content from <https://github.com/LBHackney-IT/ways-of-working>
on the 15th of October 2024 (commit <https://github.com/LBHackney-IT/ways-of-working/commit/0316206831ebecc2965f6a8cd7f407e44b1c8916>) and adds it into the root site.

This change DOES break old permalinks, in the interest of cleanliness. The original structure
had the root page for the ways of working site set as "HackIt's Development PR process" and a
child page in the site was titled "ways of working". This has been fixed so the ways of working
content is the root of the ways of working section."