Fixing image links, ordering
We can (and should) use autogenerated sidebars for ease of maintenance. In order
to migrate to that, this change adds a directory structure that mirrors the old content's
static structure.

Note: rather than linking to the same file from multiple places, which is a bit confusing,
I've just included the file once in the first pillar it was found. As a result some pillars are empty
and other have fewer contents than the previous site. This is a feature, not a bug