Migrate Product Playbook
The product playbook has been migrated from this commit: https://github.com/LBHackney-IT/Product-Playbook/commit/b14581f4f4d8eb9d4430157b0014d4d38409605f.

The Product Playbook _probably_ shouldn't live as part of the engineering site, but as the
only contributors are engineers this migration reduces the overall maintenance burden.

If this one is contentious we can revert this commit and discuss it separately