Replace absolute links with relative ones
We previously used absolute links because there was so much linking between micro-sites. Now the content lives in a single repository/build we can use relative links with impunity. This has a big advantage: Docusaurus will error on broken links in the build, so we can have confidence that internal linking works.

With this change there are zero links to playbook.hackney.gov.uk in the site content.