🧹 Replace number prefixed directories with YAML metadata
This follows the recommendation in the docs: https://docusaurus.io/docs/sidebar/autogenerated\#using-number-prefixes.

Markdown links are best used to point to the source file, i.e. including the file extension. This is then resolved to the correct path by the build
and will also work directly in the GitHub UI as a fallback and when editing there. Sadly this fails when using number prefixes, so this fix is also necessary to follow that best practice