Add Cloud Engineering to code owners
This site is growing to include ways of working that encompass development and cloud infrastructure, so it's helpful to get their review on PRs and share ownership.

This change adds Cloud Engineering as an owner across all docs. It may make more sense to scope this to just docs/architecture and docs/ways-of-working but I don't know if we care to be that fine grained right now.