Correct leads teams process
Previously these docs said that the leads only needed to be added to the '... leads_' team and permissions would be inherited for regular contributions. This is true for the repositories, however the leads also need to be able to manage the team members list, which requires them to have 'maintainer' role on the team (and also to be in that team).