⬆️ Pin cheerio to RC, upgrade Docusaurus
In August 2024 Cheerio went from RC12 to 1.0.0. This included a bunch of
breaking changes, one of which is
https://github.com/cmfcmf/docusaurus-search-local/issues/234. There
isn't a fix available for the latest version of Cheerio, but updating
dependencies pulls in the latest non-RC version automatically. To work
around this, we'll pin Cheerio to the RC until the issue is fixed.

This change enabled a full update of the dependency tree as well as
bumping Docusaurus to latest.